fix(db): widen candles.id to BigInteger to prevent PK sequence exhaustion
candles.id was int4 (max 2,147,483,647). The save path uses INSERT ... ON
CONFLICT DO NOTHING, and Postgres advances the sequence on every ATTEMPTED row
(duplicates included). The 30s crawl re-submits mostly-duplicate candles across
thousands of symbols, so the sequence burned ~2.9B values for only ~5.3M live
rows and hit the int4 ceiling on 2026-06-19 — raising
SequenceGeneratorLimitExceededError and failing EVERY candle INSERT, which froze
the price feed and starved the Intelligence Hub (no fresh candles to score).

BigInteger makes exhaustion a non-issue. This repo builds its schema from the
model via create_all (migrations/versions is gitignored), so the model change is
the durable fix — a fresh DB now creates candles.id as bigint. The live DB was
already manually altered to bigint; a matching local 0004 migration documents the
ALTER for any Alembic-managed environment.

+++ autotrade-backend/db/models.py
@@ -201,7 +201,13 @@ class Candle(Base):
         UniqueConstraint("symbol", "timeframe", "timestamp", name="uq_candle_bar"),
     )
 
-    id:        Mapped[int]      = mapped_column(Integer, primary_key=True, autoincrement=True)
+    # BigInteger, NOT Integer: candles is a very high-churn table and the save path
+    # uses INSERT ... ON CONFLICT DO NOTHING, which burns a sequence value on every
+    # ATTEMPTED row (duplicates included). An int4 PK exhausted at 2.14B and froze
+    # ALL candle writes on 2026-06-19 (SequenceGeneratorLimitExceededError), which
+    # starved the price feed + Hub. bigint makes exhaustion a non-issue.
+    # See migration 0004_candles_id_bigint.
+    id:        Mapped[int]      = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     symbol:    Mapped[str]      = mapped_column(String(50),  nullable=False)
     timeframe: Mapped[str]      = mapped_column(String(10),  nullable=False)
     open:      Mapped[float]    = mapped_column(Float, nullable=False)
